+++ controlplane/migrations/0124_fancy_greymalkin.sql
@@ -1,0 +1,7 @@
+ALTER TABLE "api_keys" DROP CONSTRAINT "api_keys_group_id_organization_groups_id_fk";
+--> statement-breakpoint
+DO $$ BEGIN
+ ALTER TABLE "api_keys" ADD CONSTRAINT "api_keys_group_id_organization_groups_id_fk" FOREIGN KEY ("group_id") REFERENCES "public"."organization_groups"("id") ON DELETE set null ON UPDATE no action;
+EXCEPTION
+ WHEN duplicate_object THEN null;
+END $$;