+++ controlplane/src/core/bufservices/cache-warmer/deleteCacheWarmerOperation.ts
@@ -5,12 +5,13 @@ import {
   DeleteCacheWarmerOperationRequest,
   DeleteCacheWarmerOperationResponse,
 } from '@wundergraph/cosmo-connect/dist/platform/v1/platform_pb';
-import { CacheWarmerRepository } from '../../../core/repositories/CacheWarmerRepository.js';
-import { FederatedGraphRepository } from '../../../core/repositories/FederatedGraphRepository.js';
-import { DefaultNamespace, NamespaceRepository } from '../../../core/repositories/NamespaceRepository.js';
-import { OrganizationRepository } from '../../../core/repositories/OrganizationRepository.js';
+import { CacheWarmerRepository } from '../../repositories/CacheWarmerRepository.js';
+import { FederatedGraphRepository } from '../../repositories/FederatedGraphRepository.js';
+import { DefaultNamespace, NamespaceRepository } from '../../repositories/NamespaceRepository.js';
+import { OrganizationRepository } from '../../repositories/OrganizationRepository.js';
 import type { RouterOptions } from '../../routes.js';
 import { enrichLogger, getLogger, handleError } from '../../util.js';
+import { UnauthorizedError } from '../../errors/errors.js';
 
 export function deleteCacheWarmerOperation(
   opts: RouterOptions,
@@ -28,13 +29,8 @@ export function deleteCacheWarmerOperation(
     const fedGraphRepo = new FederatedGraphRepository(logger, opts.db, authContext.organizationId);
     const organizationRepo = new OrganizationRepository(logger, opts.db);
 
-    if (!authContext.hasWriteAccess) {
-      return {
-        response: {
-          code: EnumStatusCode.ERR,
-          details: `The user doesnt have the permissions to perform this operation`,
-        },
-      };
+    if (authContext.organizationDeactivated || !authContext.rbac.isOrganizationAdminOrDeveloper) {
+      throw new UnauthorizedError();
     }
 
     const cacheWarmerFeature = await organizationRepo.getFeature({