+++ shared/CHANGELOG.md
@@ -4,6 +4,24 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.37.3](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.37.2...@wundergraph/cosmo-shared@0.37.3) (2025-05-26)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
+## [0.37.2](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.37.1...@wundergraph/cosmo-shared@0.37.2) (2025-05-25)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
+## [0.37.1](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.37.0...@wundergraph/cosmo-shared@0.37.1) (2025-05-24)
+
+**Note:** Version bump only for package @wundergraph/cosmo-shared
+
+# [0.37.0](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.36.17...@wundergraph/cosmo-shared@0.37.0) (2025-05-19)
+
+### Features
+
+* **router:** grpc go plugin system ([#1866](https://github.com/wundergraph/cosmo/issues/1866)) ([280a61d](https://github.com/wundergraph/cosmo/commit/280a61de4bd1328549a023d1a3a0b702d78453b8)) (@Noroth)
+
 ## [0.36.17](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-shared@0.36.16...@wundergraph/cosmo-shared@0.36.17) (2025-05-16)
 
 **Note:** Version bump only for package @wundergraph/cosmo-shared